+++ src/Resources/app/administration/src/module/sw-cms/blocks/ultimate-cms-tools/button/component/index.js
@@ -1,0 +1,5 @@
+import template from './sw-cms-block-button.html.twig';
+
+Shopware.Component.register('sw-cms-block-button', {
+    template,
+});