+++ src/Resources/app/administration/src/module/sw-cms/blocks/ultimate-cms-tools/button/index.js
@@ -1,0 +1,20 @@
+import './component';
+import './preview';
+
+Shopware.Service('cmsService').registerCmsBlock({
+    name: 'button',
+    label: 'sw-cms.blocks.ultimateCmsTools.button.label',
+    category: 'ultimate-cms-tools',
+    component: 'sw-cms-block-button',
+    previewComponent: 'sw-cms-preview-button',
+    defaultConfig: {
+        marginBottom: '20px',
+        marginTop: '20px',
+        marginLeft: '20px',
+        marginRight: '20px',
+        sizingMode: 'boxed',
+    },
+    slots: {
+        content: 'button',
+    },
+});