+++ src/Resources/app/administration/src/module/sw-cms/blocks/ultimate-cms-tools/button/preview/index.js
@@ -1,0 +1,6 @@
+import template from './sw-cms-preview-button.html.twig';
+import './sw-cms-preview-button.scss';
+
+Shopware.Component.register('sw-cms-preview-button', {
+    template,
+});