+++ src/Resources/app/administration/src/module/sw-cms/elements/button/preview/index.js
@@ -1,0 +1,8 @@
+import template from './sw-cms-el-preview-button.html.twig';
+import './sw-cms-el-preview-button.scss';
+
+const { Component } = Shopware;
+
+Component.register('sw-cms-el-preview-button', {
+    template,
+});